requirements.txt
binance-connector 
python-dotenv 
pandas
pyyaml
pyarrow
